pyproject.toml
[tool.poetry]
name = "stunning-wallpaper"
version = "0.1.0"
description = ""
authors = ["jagrutgala <jagrutgala@gmail.com>"]
readme = "README.md"
packages = [{include = "stunning_wallpaper"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.2"
pillow = "^10.0.0"
scikit-image = "^0.21.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
